Somos Project
This is the point that I can reach with my knowledge and the little time I had to study and develop due to being in the last month of current service.
I believe that with more time for both I would be able to learn what is necessary to finish the project, I apologize, and I really want to grow from that.